Move followup questions to example.
